Relocate config file
