+++ Web/services/users.ts
@@ -1,4 +1,4 @@
-import { URL_BASE_API } from '../../config';
+import { URL_BASE_API } from '../config';
 
 export const getUsers = async () => {
   const response = await fetch(`${URL_BASE_API}/users`);
